evolve kubermatic-e2e to more general purpose tool
* write out kubeconfig
* flagopts.KubeconfigFlag
* README chagens
