--- api/Dockerfile.e2e
@@ -1,34 +1,0 @@
-FROM alpine:3.7 AS builder
-
-LABEL maintainer "artiom@loodse.com"
-
-ENV KUBE_VERSION=v1.10.7
-ENV KUBE_TEST_URL=https://storage.googleapis.com/kubernetes-release/release/${KUBE_VERSION}/kubernetes-test.tar.gz
-ENV KUBE_TEST_CHECKSUM=a76614ed64c7490bca7d916f1ed98cb69cf93678
-ENV KUBECTL_URL=https://storage.googleapis.com/kubernetes-release/release/${KUBE_VERSION}/bin/linux/amd64/kubectl
-ENV KUBECTL_CHECKSUM=ab40b93f01b02657ed455b94bb9e6a73b0784501
-
-RUN apk add --no-cache curl
-RUN mkdir -p /_cache
-
-RUN curl -SsL ${KUBE_TEST_URL} -o /_cache/kubernetes-test.tar.gz
-RUN curl -SsL ${KUBECTL_URL} -o /_cache/kubectl
-RUN chmod +x /_cache/kubectl
-
-RUN echo "${KUBE_TEST_CHECKSUM}  kubernetes-test.tar.gz" > /_cache/checksum.txt
-RUN echo "${KUBECTL_CHECKSUM}  kubectl" >> /_cache/checksum.txt
-
-WORKDIR /_cache
-
-RUN sha1sum -c checksum.txt
-RUN tar xzf kubernetes-test.tar.gz
-
-####
-FROM ubuntu:16.04
-
-COPY ./_build/kubermatic-e2e /
-COPY --from=builder /_cache/kubernetes/platforms/linux/amd64/ginkgo /usr/local/bin/
-COPY --from=builder /_cache/kubernetes/platforms/linux/amd64/e2e.test /usr/local/bin/
-COPY --from=builder /_cache/kubectl /usr/local/bin/
-
-ENTRYPOINT [ "/kubermatic-e2e" ]
