+++ api/Gopkg.lock
@@ -1765,7 +1765,6 @@
     "github.com/aws/aws-sdk-go/service/ec2",
     "github.com/aws/aws-sdk-go/service/iam",
     "github.com/coreos/go-oidc",
-    "github.com/davecgh/go-spew/spew",
     "github.com/digitalocean/godo",
     "github.com/evanphx/json-patch",
     "github.com/ghodss/yaml",
@@ -1896,6 +1895,7 @@
     "k8s.io/client-go/util/cert",
     "k8s.io/client-go/util/cert/triple",
     "k8s.io/client-go/util/flowcontrol",
+    "k8s.io/client-go/util/homedir",
     "k8s.io/client-go/util/retry",
     "k8s.io/client-go/util/workqueue",
     "k8s.io/code-generator/cmd/client-gen",
